// src/app/core/models/garanties.ts
export class Garanties {
    idGarantie: number | undefined;
    idNationnal: string | undefined;
    description: string | undefined;
    valeurEnvloppe: number | undefined;
    dateDebut: Date | undefined;
    dateFin: Date | undefined;
    categorieGarantie: string | undefined;
    withTaxe: boolean | undefined;
}